Allow usersInfo to accept an authenticationDatabase
usersInfo is per database like listCollections.

An authenticationDatabase can actually be any DB on the mongod, there isn’t a separate list of authentication databases AFAICT, and you can create your users on any database, it’s just not clear to me why you’d want the configuration headache of them being outside the admin database.

Thus I believe it is clearer to label the parameter as an authenticationDatabase as that is how it behaves as you can hopefully see in my testing result expectations.

However the official MongoDB docs do at this time just say it can be performed against any “database”, with no reference to authentication on the page at all:
https://docs.mongodb.com/manual/reference/command/usersInfo/