:fire::boom: Remove the buildInfo, hostInfo, connectionStatus, usersInfo calls
These calls are probably not used downstream as they were added to try to make this refactoring from instance-model to data-service easier. If they are used downstream it is easy to add them back in. Plus connectionStatus should take an options object (for showPrivileges flag) but does not currently do so.