COMPASS-816 Move instance model node driver calls to data service (#126)
* Add TODO with my high-level intention

Obviously it needs to be made to work again, ideally with the DataService object instead of the lower-level node Db object.

* :tada: Copy/paste fetch.js from mongodb-instance-model

* npm install --save mongodb-security

* :art: var => const/let

* :tada: Inline fetch.js:get function from mongodb-instance-model

* :fire: Remove unused exports

* :art: Connect to data-service version of fetch.js

* :white_check_mark: Copy/paste fetch.test.js from mongodb-instance-model

* :fire: Remove fixtures tests

* :art: Rename instance-model to data-service

* :art: Update import and rename fetch -> getInstance

* :bug: Fix async.auto regression from v1.x to v2.x upgrade

:yay: tests.

https://github.com/caolan/async/blob/master/CHANGELOG.md#breaking-changes

* :art: Rename fetch.js to instance-detail-helper.js

* :fire::boom: Remove the buildInfo, hostInfo, connectionStatus, usersInfo calls

These calls are probably not used downstream as they were added to try to make this refactoring from instance-model to data-service easier. If they are used downstream it is easy to add them back in. Plus connectionStatus should take an options object (for showPrivileges flag) but does not currently do so.

* :art: var => const/let

* :art: More fetch -> instance renaming

* :vertical_traffic_light: Travis - Switch to port 27018

Also updates the runner syntax to that in native-client.test.js which is passing on Travis.

* :tada: Copy/paste more tests from instance model + updates

i.e. the old instance-model/test/fetched-mocked.test.js
then var/const pass, async v1 -> v2, etc.

Personal review revealed this should be paired up with the delete from instance-model when it comes to Compass, as semver allows easy management of all that.

* :shirt: Update remaining fetch -> mongodb-data-service

* :fire: Remove unused mongodb-instance-model dependency

Only Compass needs it now.